requirements.txt

---

### **3. requirements.txt**
All required libraries.

```txt
psycopg2-binary==2.9.1
openai==0.14.0
beautifulsoup4==4.10.0
requests==2.26.0
